Remove UI_CRT and other UI gunk
tiles_crt_popup() shouldn't be necessary, since these UIs don't use any
cgotoxy/cprintf calls themselves. I'm not exactly sure of the purpose of
the set_ui_state() calls here, but they shouldn't be required either for
those UIs to work correctly.

In general, we're trying to remove tiles_crt_popup() entirely, as it was
an temporary hack used to wrap old-style UIs. As of this commit, the
last remaining such UI is the skill menu UI.